requirements.txt
yapf
absl-py
numpy
pandas
seaborn
tensorflow
tensorflow_probability
tensorboardX
ipykernel